Remove redundant static_cast

+++ Samples/Desktop/D3D12Raytracing/src/D3D12RaytracingProceduralGeometry/D3D12RaytracingProceduralGeometry.cpp
@@ -1302,7 +1302,7 @@ void D3D12RaytracingProceduralGeometry::OnUpdate()
         m_animateGeometryTime += elapsedTime;
     }
     UpdateAABBPrimitiveAttributes(m_animateGeometryTime);
-    m_sceneCB->elapsedTime = static_cast<float>(m_animateGeometryTime);
+    m_sceneCB->elapsedTime = m_animateGeometryTime;
 }
